backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.20
pandas==2.2.3
pdfplumber==0.11.4
python-dateutil==2.9.0
aiofiles==24.1.0
eval-type-backport==0.3.1
